[InitializeOnLoad] not available on building exe. Put Layering.cs in Editor folder to fix
