+++ Pixel Game/Assets/Scripts/Editor/Layering.cs
@@ -1,0 +1,17 @@
+﻿using UnityEngine;
+using UnityEngine.Rendering;
+using UnityEditor;
+[InitializeOnLoad]
+class SpriteSorter
+{
+    static SpriteSorter()
+    {
+        Initialize();
+    }
+    [RuntimeInitializeOnLoadMethod]
+    static void Initialize()
+    {
+        GraphicsSettings.transparencySortMode = TransparencySortMode.CustomAxis;
+        GraphicsSettings.transparencySortAxis = new Vector3(0.0f, 1.0f, 1.0f);
+    }
+}
